+++ khoros/structures/boards.py
@@ -6,7 +6,7 @@
 :Example:           ``board_url = boards.create(khoros_object, 'my-board', 'My Board', 'forum', return_url=True)``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     05 Jun 2023
+:Modified Date:     11 Sep 2023
 """
 
 import warnings
@@ -370,7 +370,7 @@ def _structure_label_settings(_label_settings, _payload):
         else:
             # TODO: Leverage the logger instead of warnings
             warn_msg = f"The string '{_label_settings.get('allowed_labels')}' for the 'allowed_labels' field is " \
-                       f"not valid and will be ignored."
+                       "not valid and will be ignored."
             warnings.warn(warn_msg, UserWarning)
             _label_settings['allowed_labels'] = None
 
@@ -379,8 +379,8 @@ def _structure_label_settings(_label_settings, _payload):
             _label_settings.get('allowed_labels') is not None or
             _label_settings.get('allowed_labels') is not None):
         # TODO: Leverage the logger instead of warnings
-        warn_msg = f"The defined 'allowed_labels' field will be overwritten when the 'use_freeform_labels' and/or " \
-                   f"'use_predefined_labels' Boolean values are also configured."
+        warn_msg = "The defined 'allowed_labels' field will be overwritten when the 'use_freeform_labels' and/or " \
+                   "'use_predefined_labels' Boolean values are also configured."
         warnings.warn(warn_msg, UserWarning)
 
     # Define the 'allowed_labels' value based on the defined Boolean values when applicable
@@ -528,8 +528,11 @@ def get_message_count(khoros_object, board_id):
     return message_count
 
 
-def get_all_messages(khoros_object, board_id, fields=None):
+def get_all_messages(khoros_object, board_id, fields=None, where_filter=None, descending=True):
     """This function retrieves data for all messages within a given board.
+
+    .. versionchanged:: 5.4.0
+       Introduced the ``where_filter`` and ``descending`` parameters to more specifically adjust the LiQL query.
 
     .. versionadded:: 5.3.0
 
@@ -539,6 +542,10 @@ def get_all_messages(khoros_object, board_id, fields=None):
     :type board_id: str
     :param fields: Specific fields to query if not all fields are needed (comma-separated string or iterable)
     :type fields: str, tuple, list, set, None
+    :param where_filter: One or more optional WHERE filters to include in the LiQL query
+    :type where_filter: str, tuple, list, set, None
+    :param descending: Determines if the data should be returned in descending order (``True`` by default)
+    :type descending: bool
     :returns: A list containing a dictionary of data for each message within the board
     :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
     """
@@ -549,7 +556,10 @@ def get_all_messages(khoros_object, board_id, fields=None):
     fields = '*' if not fields else fields
     if not isinstance(fields, str):
         fields = liql.parse_select_fields(fields)
-    query = f"SELECT {fields} FROM messages WHERE board.id = '{board_id}' ORDER BY last_publish_time DESC LIMIT 1000"
+    where_clause = _construct_where_clause(where_filter)
+    order_by_clause = _construct_order_by_clause(where_filter, descending)
+    query = f"SELECT {fields} FROM messages WHERE board.id = '{board_id}' " \
+            f"{where_clause}ORDER BY {order_by_clause} LIMIT 1000"
 
     # Perform the first LiQL query and add to the master list
     response, cursor = _perform_single_query(khoros_object, query, fields)
@@ -562,6 +572,61 @@ def get_all_messages(khoros_object, board_id, fields=None):
 
     # Return the collected messages
     return messages
+
+
+def get_all_topic_messages(khoros_object, board_id, fields=None, descending=True):
+    """This function retrieves data for all topic messages (i.e. zero-depth messages) within a given board.
+
+    .. versionadded:: 5.4.0
+
+    :param khoros_object: The core :py:class:`khoros.Khoros` object
+    :type khoros_object: class[khoros.Khoros]
+    :param board_id: The ID of the board to query
+    :type board_id: str
+    :param fields: Specific fields to query if not all fields are needed (comma-separated string or iterable)
+    :type fields: str, tuple, list, set, None
+    :param descending: Determines if the data should be returned in descending order (``True`` by default)
+    :type descending: bool
+    :returns: A list containing a dictionary of data for each topic message within the board
+    :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
+    """
+    return get_all_messages(khoros_object, board_id, fields, where_filter='depth=0', descending=descending)
+
+
+def _construct_where_clause(_where_filter=None):
+    """This function constructs a supplemental WHERE clause from any filters that are provided.
+
+    .. versionadded:: 5.4.0
+
+    :param _where_filter: Zero or more WHERE filters
+    :type _where_filter: str, list, set, tuple, None
+    :returns: The constructed supplemental WHERE clause as a string
+    """
+    _where_clause = ''
+    if _where_filter:
+        _where_filter = [_where_filter] if isinstance(_where_filter, str) else _where_filter
+        for _filter in _where_filter:
+            if _filter:
+                _where_clause += f'AND {_filter} '
+    return _where_clause
+
+
+def _construct_order_by_clause(_where_filter=None, _descending=True):
+    """This function constructs the ORDER BY clause for the :py:func:`khoros.objects.boards.get_all_messages` function.
+
+    .. versionadded:: 5.4.0
+
+    :param _where_filter: Zero or more WHERE filters
+    :type _where_filter: str, list, set, tuple, None
+    :param _descending: Determines if the data should be returned in descending order (``True`` by default)
+    :type _descending: bool
+    :returns: The constructed ORDER BY clause
+    """
+    _field = 'last_publish_time'
+    if isinstance(_where_filter, str) and 'depth=0' in _where_filter:
+        _field = 'conversation.last_post_time'
+    _direction = 'DESC' if _descending else 'ASC'
+    return f'{_field} {_direction}'
 
 
 def _perform_single_query(khoros_object, query, fields=None, cursor=None):